Added loading previous versions typings on website build
