+++ .gitignore
@@ -26,7 +26,7 @@ debug.html
 /website/build
 /website/.docusaurus
 /website/.cache-loader
-/website/.cache/
+/website/.previous-typings-cache/*.d.ts
 
 !/src/typings/**/*.d.ts
