+++ website/.previous-typings-cache/tsconfig.json
@@ -1,0 +1,6 @@
+{
+	"include": [
+		"./*.d.ts",
+		"../../dist/typings.d.ts"
+	]
+}
